requirements.txt
streamlit==1.39.0
google-cloud-aiplatform
pandas
numpy
python-dotenv
streamlit-js-eval
st-pages==0.5.0
PyMuPDF
streamlit_pdf_viewer==0.0.12
ipython
google-cloud-logging
scikit-learn
langchain-pinecone
langchain 
langchain-community
langchain-experimental 
xmltodict
pubmed-api
rapidocr-onnxruntime
pymupdf4llm
langchain-google-genai
pypdf
pinecone-client
semanticscholar
langchainhub
lark
langchain-chroma
google-ai-generativelanguage
langchainhub 
google-generativeai
langchain-google-community
beautifulsoup4
html2text
faiss-cpu
streamlit-authenticator
rapidocr-onnxruntime
pymupdf4llm